libs/shared/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "fraud-shared"
version = "0.1.0"
description = "Shared libraries for the Fraud Decision Support Platform"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "pydantic>=2.5",
  "pydantic-settings>=2.1",
  "prometheus-client>=0.19",
  "redis>=5.0",
  "asyncpg>=0.29",
  "PyJWT>=2.8",
  "opentelemetry-api>=1.24",
  "opentelemetry-sdk>=1.24"
]

[tool.setuptools]
package-dir = {"" = "shared"}

[tool.setuptools.packages.find]
where = ["shared"]
